Rename files and reorder imports
